delete grade_check_layout and GradeCheckFragment
