--- KuleumBridge/app/src/main/res/layout/grade_check_layout.xml
@@ -1,104 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:background="@drawable/background1"
-    android:orientation="vertical"
-    android:gravity="center">
-
-    <!-- 성적 조회 레이아웃 -->
-    <LinearLayout
-        android:id="@+id/grade_check_layout_sep"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:layout_gravity="center"
-        android:orientation="vertical"
-        android:visibility="visible">
-
-
-        <TableLayout
-            android:id="@+id/grade_now_tablelayout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="100dp"
-            android:background="@drawable/rounded_corner"
-            android:isScrollContainer="true"
-            android:stretchColumns="*">
-
-            <TableRow
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="20dp">
-
-
-                <TextView
-
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginLeft="5dp"
-                    android:layout_weight="1.0"
-                    android:gravity="center"
-                    android:text="이수구분"
-                    android:textSize="20sp"
-                    android:textStyle="bold"
-                    android:textColor="@color/black"/>
-
-                <TextView
-
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-
-                    android:layout_weight="2.0"
-                    android:gravity="center"
-                    android:text="과목명"
-                    android:textSize="18sp"
-                    android:textStyle="bold"
-                    android:textColor="@color/black"/>
-
-                <TextView
-
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="0.5"
-                    android:gravity="center"
-                    android:text="학점"
-                    android:textSize="18sp"
-                    android:textStyle="bold"
-                    android:textColor="@color/black"/>
-
-                <TextView
-
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginRight="5dp"
-                    android:layout_weight="0.7"
-                    android:gravity="center"
-                    android:text="성적"
-                    android:textSize="18sp"
-                    android:textStyle="bold"
-                    android:textColor="@color/black"/>
-            </TableRow>
-
-
-        </TableLayout>
-
-
-        <Button
-            android:id="@+id/grade_all_check_button"
-            android:layout_width="150dp"
-            android:layout_height="50dp"
-            android:layout_gravity="center"
-            android:layout_marginTop="30dp"
-            android:backgroundTint="@color/black"
-            android:onClick="onClick"
-            android:gravity="center"
-            android:text="↗전체 성적 조회 "
-            android:textSize="16sp"
-            android:textColor="@color/white"
-            android:textStyle="bold" />
-    </LinearLayout>
-
-</LinearLayout>
